refactor: fixing folder structure
