+++ SensorApp/SensorApp.Infrastructure/Api/Dtos/AuthResponseDto.cs
@@ -1,0 +1,8 @@
+﻿namespace SensorApp.Infrastructure.Api.Dtos;
+
+internal class AuthResponseDto
+{
+    public string UserId { get; set; }
+    public string Username { get; set; }
+    public string Token { get; set; }
+}
