+++ SensorApp/SensorApp.Infrastructure/Api/Dtos/LoginDto.cs
@@ -1,0 +1,7 @@
+﻿namespace SensorApp.Infrastructure.Api.Dtos;
+
+internal class LoginDto
+{
+    public string Username { get; set; }
+    public string Password { get; set; }
+}
