+++ SensorApp/SensorApp.Infrastructure/Api/Dtos/UserWithRoleDto.cs
@@ -1,0 +1,9 @@
+﻿namespace SensorApp.Infrastructure.Api.Dtos;
+
+internal class UserWithRoleDto
+{
+    public string Id { get; set; }
+    public string Username { get; set; }
+    public string Email { get; set; }
+    public string Role { get; set; }
+}
