+++ SensorApp/SensorApp.Infrastructure/Services/Auth/TokenService.cs
@@ -1,9 +1,10 @@
 ﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.IdentityModel.Tokens;
-using SensorApp.Infrastructure.Services.Auth;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
+
+namespace SensorApp.Infrastructure.Services.Auth;
 
 public class TokenService : ITokenService
 {
@@ -15,15 +16,15 @@ public class TokenService : ITokenService
 
     public string GenerateToken(IdentityUser user, IList<string> roles, IList<Claim> additionalClaims)
     {
-        var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_configuration["JwtSettings:Key"]));
+        var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_configuration["JwtSettings:Key"]!));
         var credentials = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
 
         var tokenClaims = new List<Claim>
         {
-            new Claim(JwtRegisteredClaimNames.Sub, user.Id),
-            new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
-            new Claim(ClaimTypes.Email, user.Email),
-            new Claim("email_confirmed", user.EmailConfirmed.ToString())
+            new(JwtRegisteredClaimNames.Sub, user.Id),
+            new (JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
+            new (ClaimTypes.Email, user.Email!),
+            new ("email_confirmed", user.EmailConfirmed.ToString())
         };
 
         tokenClaims.AddRange(additionalClaims);
